Fix compile & run error Chapter03 requests_mock
